requirements.txt
torchvision-0.14.1
numpy==1.24.4
opencv-python==4.8.1.78
aiogram==2.25.1
pandas==2.0.3
python-dotenv==1.0.0
matplotlib
scikit-learn
tqdm
torchmetrics